New and more powerful (and correct) find_var_blocks
Broke `find_var_blocks` and supporting code into its own script and reworked it to be able to handle multiple nested var blocks. This allows inner eval blocks which are not resolved (see `merge_function_unresolved_var` parser scenario).

Fixes: #816
